v0.1.2
Menu and note overflow defined
