+++ simple-note-react/src/components/IndexItem.js
@@ -2,11 +2,13 @@ import React from "react"
 
 function IndexItem(props) {
     return (
-        <div className="index-item" style={props.index !== 1 ? {borderTop: "1px solid black"} : {border: ""}}>
+        <div className="index-item" >
             <h4>{props.header}</h4>
             <p>{props.text}</p>
         </div>
     )
 }
+
+// style={props.index !== 1 ? {borderTop: "1px solid black"} : {border: ""}}
 
 export default IndexItem
