--- simple-note-react/src/components/NavBar.js
@@ -1,13 +1,0 @@
-import React, { useState } from "react"
-
-function NavBar() {
-    return (
-        <div className="navbar">
-            <div className="navbar-logo">
-                <p>SimpleNote</p>
-            </div>
-        </div>
-    )
-}
-
-export default NavBar
